dashboard/tests/unit/test_dashboard_stack.py
import aws_cdk as core
import aws_cdk.assertions as assertions
from unittest.mock import patch, MagicMock

from dashboard.dashboard_stack import DashboardStack
from dashboard.cognito_utils import get_existing_cognito_resources

# example tests. To run these tests, uncomment this file along with the example
# resource in dashboard/dashboard_stack.py
def test_get_existing_cognito_resources():
    mock_cognito_idp = MagicMock()
    mock_cognito_identity = MagicMock()

    # Test scenario 1: User Pool and Identity Pool exist
    mock_cognito_idp.list_user_pools.return_value = {
        'UserPools': [
            {'Id': 'existing_user_pool_id', 'Name': 'CognitoUserPool'}
        ]
    }
    mock_cognito_identity.list_identity_pools.return_value = {
        'IdentityPools': [
            {'IdentityPoolId': 'existing_identity_pool_id', 'IdentityPoolName': 'CognitoIdentityPool'}
        ]
    }

    with patch('dashboard.cognito_utils.boto3.client') as mock_boto3_client:
        mock_boto3_client.side_effect = [mock_cognito_idp, mock_cognito_identity]
        user_pool_id, identity_pool_id = get_existing_cognito_resources('CognitoUserPool', 'CognitoIdentityPool')

    assert user_pool_id == 'existing_user_pool_id'
    assert identity_pool_id == 'existing_identity_pool_id'

    # Test scenario 2: User Pool and Identity Pool do not exist
    mock_cognito_idp.list_user_pools.return_value = {'UserPools': []}
    mock_cognito_identity.list_identity_pools.return_value = {'IdentityPools': []}

    with patch('dashboard.cognito_utils.boto3.client') as mock_boto3_client:
        mock_boto3_client.side_effect = [mock_cognito_idp, mock_cognito_identity]
        user_pool_id, identity_pool_id = get_existing_cognito_resources('NonExistentPool', 'NonExistentPool')

    assert user_pool_id is None
    assert identity_pool_id is None


def test_dashboard_stack():
    app = core.App()
    stack = DashboardStack(app, "TestDashboardStack")
    template = assertions.Template.from_stack(stack)

    # Add assertions for the expected resources and properties
    # ...

