dashboard/requirements.txt
aws-cdk-lib==2.94.0
constructs>=10.0.0,<11.0.0